agentic-city-pulse-backend-server-main/agents/vertex_ai/requirements_vertex_ai.txt
# Core dependencies
streamlit>=1.31.0
python-dotenv>=1.0.0

# Google Cloud and Vertex AI
google-cloud-aiplatform>=1.38.0
google-cloud-firestore>=2.11.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Firebase
firebase-admin>=6.2.0

# Vertex AI specific
vertexai>=1.38.0

# RAG and Vector Search dependencies
torch>=1.11.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0

# Additional utilities
requests>=2.31.0
pandas>=2.0.0 